services/export-service/requirements.txt
# AutoGraph v3 - Export Service Dependencies

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0

# Image processing
Pillow==11.0.0

# PDF generation
reportlab==4.2.5

# Utilities
python-dotenv==1.0.1
